[SPARK-36010][BUILD] Upgrade sbt-antlr4 from 0.8.2 to 0.8.3
### What changes were proposed in this pull request?

This PR proposes to upgrade sbt-antlr4 from 0.8.2 to 0.8.3 per the guides at https://github.com/ihji/sbt-antlr4
I can't find an official proper docs for this.

### Why are the changes needed?

To stick to the guides in https://github.com/ihji/sbt-antlr4, and leverage the fixes included.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

CI in this PR should tests it out.

Closes #33208 from HyukjinKwon/SPARK-36010.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Kousuke Saruta <sarutak@oss.nttdata.com>